backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "自动化改作业助教 Agent 后端服务"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.136.0",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "python-multipart>=0.0.26",
    "rapidfuzz>=3.14.5",
    "rapidocr-onnxruntime>=1.4.4",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]

[project.scripts]
backend = "backend:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
